+++ src/strojure/memoize_one/core.clj
@@ -1,6 +1,5 @@
 (ns strojure.memoize-one.core
-  "Memoization of the single value."
-  (:import (clojure.lang IDeref)))
+  "Memoization of the single value.")
 
 (set! *warn-on-reflection* true)
 
@@ -10,7 +9,7 @@
   "Functions for the cached computation."
 
   (get-ref
-    ^IDeref [cache]
+    ^clojure.lang.IDeref [cache]
     "Returns cached value reference to `deref`."))
 
 (defprotocol MemoizedRef
@@ -21,6 +20,13 @@
     "Invalidates this reference in the cache. Returns nil."))
 
 ;;••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••
+
+(defn get-value
+  "Returns result of `(deref (get-ref cache))`."
+  {:inline (fn [cache] `(.deref (get-ref ~cache)))
+   :added "1.1"}
+  [cache]
+  (.deref (get-ref cache)))
 
 (defn get-another-ref
   "Invalidates memoized ref and returns actual (probably new) ref from the cache."