+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 Release date `UNRELEASED`
 
 - Add typehint for return value of the `core/get-ref`.
+- New API function `core/get-value`.
 
 ## `1.0.3`